+++ ZeroLibraries/Serialization/DataTree.cpp
@@ -376,14 +376,14 @@ void DataTreeLoader::PushChildOnStack()
 
 //---------------------------------------------------------- Array Serialization 
 //******************************************************************************
-void DataTreeLoader::ArraySize(uint& arraySize)
+void DataTreeLoader::ArraySize(size_t& arraySize)
 {
   arraySize = GetCurrent()->GetNumberOfChildren();
 }
 
 //******************************************************************************
 template<typename type>
-void ReadArray(DataNode* arrayNode, type* data, uint numberOfElements)
+void ReadArray(DataNode* arrayNode, type* data, size_t numberOfElements)
 {
   ErrorIf(arrayNode->mNodeType != DataNodeType::Object, "Node is not an array type. "
           "Name:'%s'", arrayNode->mPropertyName.c_str());
@@ -392,7 +392,7 @@ void ReadArray(DataNode* arrayNode, type* data, uint numberOfElements)
           arrayNode->GetNumberOfChildren());
 
   DataNode* curElem = &arrayNode->GetChildren().Front();
-  for(uint i = 0; i < numberOfElements && curElem; ++i)
+  for(size_t i = 0; i < numberOfElements && curElem; ++i)
   {
     DataNode* lit = curElem;
     ToValue(lit->mTextValue.All(), data[i]);
@@ -402,7 +402,7 @@ void ReadArray(DataNode* arrayNode, type* data, uint numberOfElements)
 
 //******************************************************************************
 bool DataTreeLoader::ArrayField(cstr typeName, cstr fieldName, byte* data, ArrayType arrayType, 
-                                uint numberOfElements, uint sizeOftype)
+                                size_t numberOfElements, size_t sizeOftype)
 {
   if(InnerStart(typeName, fieldName, StructureType::BasicArray))
   {
